BackEnd/requirements.txt
psycopg2
sqlalchemy
python-dotenv

